cart

+++ shop/products/models.py
@@ -5,7 +5,7 @@ from datetime import datetime
 class Addproduct(db.Model):
     id = db.Column(db.Integer, primary_key=True,)
     name = db.Column(db.String(80), nullable=False)
-    price= db.Column(db.Integer, nullable= False)
+    price= db.Column(db.Numeric(10,2), nullable= False)
     discount = db.Column(db.Integer, default=0)
     stock = db.Column(db.Integer, nullable = False)
